+++ src/tools/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Barrel export for all Misty bot tools.
+ * Tools are organized by functionality and use factory functions
+ * to inject dependencies like Discord messages and client state.
+ */
+
+export { createMyselfTool } from "./myself.js";
+export { createSendMessageTool } from "./sendMessage.js";
+export { createPlayMusicTool } from "./playMusic.js";
+export { createStopPlayingTool } from "./stopPlaying.js";
+export { createWhatSongTool } from "./whatSong.js";
